+++ libevmasm/SimplificationRule.h
@@ -136,6 +136,8 @@ struct EVMBuiltins
 	static auto constexpr MSTORE8 = PatternGenerator<Instruction::MSTORE8>{};
 	static auto constexpr SLOAD = PatternGenerator<Instruction::SLOAD>{};
 	static auto constexpr SSTORE = PatternGenerator<Instruction::SSTORE>{};
+	static auto constexpr TLOAD = PatternGenerator<Instruction::TLOAD>{};
+	static auto constexpr TSTORE = PatternGenerator<Instruction::TSTORE>{};
 	static auto constexpr PC = PatternGenerator<Instruction::PC>{};
 	static auto constexpr MSIZE = PatternGenerator<Instruction::MSIZE>{};
 	static auto constexpr GAS = PatternGenerator<Instruction::GAS>{};