+++ test/libsolidity/semanticTests/inlineAssembly/transient_storage_creation.sol
@@ -1,0 +1,14 @@
+contract C {
+    constructor() {
+        uint x;
+        assembly {
+            tstore(0, 42)
+            x := tload(0)
+        }
+        assert(x == 42);
+    }
+}
+// ====
+// EVMVersion: >=cancun
+// ----
+// constructor() ->