+++ test/libsolidity/semanticTests/inlineAssembly/transient_storage_selfdestruct.sol
@@ -1,0 +1,44 @@
+contract C {
+    function set(uint value) external {
+        assembly {
+            tstore(0, value)
+        }
+    }
+
+    function get() external view returns (uint value) {
+        assembly {
+            value := tload(0)
+        }
+    }
+
+    function terminate(address payable a) external {
+        selfdestruct(a);
+    }
+}
+
+contract D {
+    C public c;
+
+    constructor() {
+        c = new C();
+    }
+
+    function destroy() external {
+        c.set(42);
+        c.terminate(payable(address(this)));
+        assert(c.get() == 42);
+    }
+
+    function createAndDestroy() external {
+        c = new C();
+        c.set(42);
+        c.terminate(payable(address(this)));
+        assert(c.get() == 42);
+    }
+}
+// ====
+// EVMVersion: >=cancun
+// ----
+// constructor() ->
+// destroy() ->
+// createAndDestroy() ->