+++ Changelog.md
@@ -6,10 +6,13 @@ Language Features:
  * Yul: Introduce builtin ``blobbasefee()`` for retrieving the blob base fee of the current block.
  * Yul: Introduce builtin ``blobhash()`` for retrieving versioned hashes of blobs associated with the transaction.
  * Yul: Introduce builtin ``mcopy()`` for cheaply copying data between memory areas.
+ * Yul: Introduce builtins ``tload()`` and ``tstore()`` for transient storage access.
+
 
 Compiler Features:
 * EVM: Support for the EVM Version "Cancun".
 * SMTChecker: Support `bytes.concat` except when string literals are passed as arguments.
+
 
 Bugfixes:
  * AST import: Fix bug when importing inline assembly with empty ``let`` variable declaration.